Added databas (Contacts) table CRUDE operations - minor error needs fixing
