+++ Database/database.json
@@ -1,1 +1,1 @@
-{"Users": {"1": {"user_name": "munachi", "email": "email", "password": "12345"}, "2": {"user_name": "davido", "email": "davido@yahoomail.com", "password": "frog_voice_101"}}, "Einstein_Munachiso_contact_db": {"1": {"first_name": "Einstein", "last_name": "Munachiso", "number": "08141822937", "category": "nil", "email": "nil"}}, "Caleb_Ogudu_contact_db": {"1": {"first_name": "Caleb", "last_name": "Ogudu", "number": "08141823859", "category": "nil", "email": "nil"}}}
+{"Users": {"1": {"user_name": "munachi", "email": "email", "password": "12345"}, "2": {"user_name": "davido", "email": "davido@yahoomail.com", "password": "frog_voice_101"}}, "Einstein_Munachiso_contact_db": {"1": {"first_name": "Einstein", "last_name": "Munachiso", "number": "08141822937", "category": "nil", "email": "nil"}}, "Caleb_Ogudu_contact_db": {"1": {"first_name": "Caleb", "last_name": "Ogudu", "number": "08141823859", "category": "nil", "email": "nil"}}, "Daniel_Anchave_contact_db": {"1": {"user_id": 1, "first_name": "Daniel", "last_name": "Anchave", "number": "090274663527", "category": "Family", "email": "nil"}}}
