+++ app.py
@@ -1,4 +1,4 @@
-from database import *
+from database_control import *
 from flask import Flask, render_template, request, redirect, url_for, jsonify
 
 app = Flask(__name__)
